ride-sharing-backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-dotenv==1.0.0
requests==2.31.0
redis==4.6.0
pytest==7.4.3
pytest-django==4.5.2
gunicorn==21.2.0
psycopg2-binary==2.9.9  
geopy==2.3.0  
pytest==8.2.2
pytest-django==4.8.0
pytest-cov==4.1.0